fix(db): resolve Prisma 7 migration issues causing build failures
Prisma 7 requires driver adapters which necessitated architectural changes:

- Add @prisma/adapter-pg with PrismaPg adapter in lib/prisma.ts
- Configure serverExternalPackages in next.config.mjs for pg module
- Update all database scripts to use shared Prisma instance instead of
  creating isolated PrismaClient instances (which fail without adapter)
- Split utils/onboarding.ts into client-safe and server-only modules:
  - utils/onboarding.ts: Client-safe utilities (no DB operations)
  - utils/onboarding-server.ts: Server-only DB operations
- Update imports in onboarding.action.ts and API route
- Add comprehensive migration documentation in docs/prisma-7-migration.md

Scripts updated:
- cleanup-invalid-appointments.ts
- release-earnings.ts
- cleanup-abandoned-payments.ts
- create-payout-batch.ts
- process-payouts.ts
- appointmentUtils.ts
