+++ package.json
@@ -115,6 +115,7 @@
     "react-hook-form": "^7.69.0",
     "react-icons": "^5.5.0",
     "resend": "^6.6.0",
+    "server-only": "^0.0.1",
     "sharp": "^0.34.5",
     "stream-chat": "^9.28.0",
     "stream-chat-react": "^13.13.1",
